requirements.txt
fastapi==0.115.11
uvicorn[standard]==0.34.0
sqlalchemy==2.0.39
alembic==1.15.1
psycopg[binary]==3.2.6
pydantic==2.10.6
redis==5.2.1
celery==5.4.0
jinja2==3.1.4
python-multipart==0.0.20
pymupdf==1.26.4
python-docx==1.1.2


